requirements.txt
atproto>=0.0.60
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
mf2py>=1.1.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-flask>=1.2.0
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.0.0
gunicorn>=21.0.0
coverage>=7.0.0